The reacher environment on UR5 can now be used with higher speeds and accelerations without incurring protective stops. (#69)
* The reacher environment on UR5 can now be used with higher speeds and
accelerations without incurring protective stops.
The UR5 robot expects actuation commands every 8ms. This was violated in
the original environment and communicator, causing protective stops.
It was fixed by using the stopj command in a few places.
In addition, the original environment would change the acceleration
value permanently if the robot went out of bounds. It is now reset to
the initial value for every actuation.
These improvements were imported from
https://github.com/homayoonfarrahi/cycle-time-study/tree/master/senseact_mod.

* Removed some unnecessary lines from the UR5 environment.

---------

Co-authored-by: Homayoon Farrahi <farrahi@ualbera.ca>